+++ jib-gradle-plugin/src/test/java/com/google/cloud/tools/jib/gradle/GradleRawConfigurationTest.java
@@ -47,7 +47,6 @@ public class GradleRawConfigurationTest {
     Mockito.when(jibExtension.getTo()).thenReturn(targetImageParameters);
     Mockito.when(jibExtension.getContainer()).thenReturn(containerParameters);
     Mockito.when(jibExtension.getAllowInsecureRegistries()).thenReturn(true);
-    Mockito.when(jibExtension.getUseOnlyProjectCache()).thenReturn(true);
 
     Mockito.when(baseImageParameters.getCredHelper()).thenReturn("gcr");
     Mockito.when(baseImageParameters.getImage()).thenReturn("openjdk:15");
@@ -94,7 +93,6 @@ public class GradleRawConfigurationTest {
     Assert.assertEquals(
         new HashSet<>(Arrays.asList("additional", "tags")), rawConfiguration.getToTags());
     Assert.assertTrue(rawConfiguration.getUseCurrentTimestamp());
-    Assert.assertTrue(rawConfiguration.getUseOnlyProjectCache());
     Assert.assertEquals("admin:wheel", rawConfiguration.getUser().get());
   }
 }