+++ jib-maven-plugin/src/main/java/com/google/cloud/tools/jib/maven/MavenRawConfiguration.java
@@ -138,11 +138,6 @@ class MavenRawConfiguration implements RawConfiguration {
   }
 
   @Override
-  public boolean getUseOnlyProjectCache() {
-    return jibPluginConfiguration.getUseOnlyProjectCache();
-  }
-
-  @Override
   public ImageFormat getImageFormat() {
     return ImageFormat.valueOf(jibPluginConfiguration.getFormat());
   }